@@ -1,16 +1,16 @@
 package br.com.blz.testjava;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
-@RunWith(SpringRunner.class)
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
 @SpringBootTest
-public class TestJavaApplicationTests {
+class TestJavaApplicationTests {
 
 	@Test
-	public void contextLoads() {
+	void contextLoads() {
+		assertTrue(true);
 	}
 
 }
